+++ kmm/src/commonMain/kotlin/fr/beapp/interviews/bicloo/kmm/data/contract/ContractDataSource.kt
@@ -1,16 +1,15 @@
 package fr.beapp.interviews.bicloo.kmm.data.contract
 
-import fr.beapp.interviews.bicloo.kmm.core.CoreHelper
 import fr.beapp.interviews.bicloo.kmm.core.rest.RestClient
 import kotlinx.serialization.builtins.ListSerializer
 
 internal class ContractDataSource(private val restClient: RestClient) {
 
-	private val baseUrl = "${CoreHelper.environmentInfo.urlHost}/contracts"
+	private val basePath = "/contracts"
 
 	suspend fun getContracts(): List<ContractDTO> {
 		return restClient.get(
-			path = baseUrl,
+			path = basePath,
 			serializer = ListSerializer(ContractDTO.serializer()),
 			isAuthRequired = false,
 		)