+++ kmm/src/commonMain/kotlin/fr/beapp/interviews/bicloo/kmm/presentation/Dispatchers.kt
@@ -1,0 +1,7 @@
+package fr.beapp.interviews.bicloo.kmm.presentation
+
+import kotlin.coroutines.CoroutineContext
+
+expect val defaultDispatcher: CoroutineContext
+
+expect val uiDispatcher: CoroutineContext