+++ kmm/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `ContractEntity`, `ContractManager` and `ContractPresenter` classes
 - `StationEntity`,`StationManager` and `StationPresenter` classes
+- `ServiceLocator` object
 
 ### Chore